requirements.txt
openai
python-dotenv
praw
requests
pandas
gTTS
tqdm
scikit-learn
selenium
ffmpeg-python
Pillow
gensim
